requirements.txt
aiohttp==3.8.6
discord.py==2.3.2
loguru==0.7.2
python-dotenv==1.0.0
Requests==2.31.0